* master:
  Minor changes to resolve Visual Studio warnings
  moved public release notes in 1.0.0
  moved public release notes in v0.7.0

+++ maptk/match_matrix.cxx
@@ -54,7 +54,7 @@ match_matrix(const vital::track_set_sptr tracks,
     std::set<vital::frame_id_t> frame_ids = tracks->all_frame_ids();
     frames = std::vector<vital::frame_id_t>(frame_ids.begin(), frame_ids.end());
   }
-  const size_t num_frames = frames.size();
+  const unsigned int num_frames = static_cast<unsigned int>(frames.size());
 
   // build a frame map for reverse lookup of matrix indices
   std::map<vital::frame_id_t, unsigned int> frame_map;
@@ -65,13 +65,13 @@ match_matrix(const vital::track_set_sptr tracks,
 
   // compute an upper bound on non-zero matrix entries to
   // pre-allocate the sparse matrix memory
-  size_t max_size = 0;
+  unsigned int max_size = 0;
   const std::vector<vital::track_sptr> trks = tracks->tracks();
   VITAL_FOREACH(const vital::track_sptr& t, trks)
   {
     if( t->size() > max_size )
     {
-      max_size = t->size();
+      max_size = static_cast<unsigned int>(t->size());
     }
   }
   Eigen::SparseMatrix<unsigned int> mm(num_frames, num_frames);
